introduce the post request to push a new game

+++ src/main/java/pfr/backgamesloc/games/DAL/entities/Editor.java
@@ -3,25 +3,24 @@ package pfr.backgamesloc.games.DAL.entities;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import jakarta.persistence.*;
 import lombok.Data;
-import pfr.backgamesloc.games.DAL.entities.Game;
 
 import java.util.List;
 
 @Entity
-@Table(name ="editor")
+@Table(name = "editor")
 @Data
 public class Editor {
 
     @Id
     @Column(name = "editor_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private short editorId;
+    private Integer editorId;
 
     @Column(name = "name")
     private String name;
 
     @JsonBackReference
-    @OneToMany(fetch = FetchType.LAZY,mappedBy = "editor")
+    @OneToMany(fetch = FetchType.LAZY, mappedBy = "editor")
     private List<Game> games;
 
 }
